+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+* Added update overlay size method when overlay is displayed and added close when user tap in back button as new param `closeWhenTapBackButton` and added support for horizontal and vertical alignment
+
+## 0.0.2
+
+* Added overlay orientation as new param `screenOrientation`
+
 ## 0.0.1
 
 * Initial release